Add block methods
